WIP: Fix linters, Integrate BookManager, MusicAlbumManager, and GameManager into ConsoleApp
